Change Descendent to Manifestation
An explanation here so people understand what's going on when they go through my translations. The word translated here is "janna", which in traditional translation is interpreted as gardens or paradise. This makes the story kinda flow, but there is no other motivation for that other than the appearance here together with the fruits and the branches. The actual etymology of the word however hints to some relation to the Jinn, which is also something that appears here and is for now translated as genies. The meaning of the jinn in this text is becoming more clear to me above just a ghost-like being, and I think the meaning of janna is basically some kind of jinn, reality, sent down by God. As such I am also considering changing the translation of jinn as genie, which is just a calque of the Arabic and doesn't have any meaning in it of itself, to something like a presence or spectre, as in a figure that is present or generally a more common english presentation which is not necessarily positive or negative of something in the invisible world. It doesn't really change the understanding of a genie, but it does make the link between the presence/spectre and the Manifestation more clear.